requirements.txt
psycopg2-binary>=2.9.9
lxml>=5.0.0
python-dateutil>=2.8.2
pytest>=7.0.0
jsonschema>=4.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
ipaddress>=1.0.23
